Enable FreeRTOS library to build for Marlin
The actual configuration of the FreeRTOS library is controlled by the user's application.   In this case, it is Marlin that controls how the library is built.
This file should be moved to ../Marlin/src/inc   but the normal BUILD_FLAGS are not used when building the libraries.   Only the include_dir variable is used and right now it points at ../Marlin